[major] Connect app desired pH with pH controller on pi
